Error message for no ycmd submodule checkout

+++ install.sh
@@ -2,7 +2,15 @@
 
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-$SCRIPT_DIR/third_party/ycmd/build.sh "$@"
+build_file=$SCRIPT_DIR/third_party/ycmd/build.sh
+
+if [[ ! -f "$build_file" ]]; then
+  echo "File $build_file doesn't exist; you probably forgot to run:"
+  printf "\n\tgit submodule update --init --recursive\n\n"
+  exit 1
+fi
+
+$build_file "$@"
 
 # Remove old YCM libs if present so that YCM can start.
 rm -f python/*ycm_core.* &> /dev/null
